requirements.txt
lxml
PyYAML
pyaml
redis
requests
